Improve code readability by splitting runtime and editor parts
