+++ UOP1_Project/Assets/Scripts/SceneManagement/ScriptableObjects/GameSceneSO.cs
@@ -1,20 +1,14 @@
 ﻿using UnityEngine;
-#if UNITY_EDITOR
-using UnityEditor;
-#endif
 
 /// <summary>
 /// This class is a base class which contains what is common to all game scenes (Locations or Menus)
 /// </summary>
 
-public abstract class GameSceneSO : ScriptableObject
-#if UNITY_EDITOR
-	, ISerializationCallbackReceiver
-#endif
+public abstract partial class GameSceneSO : ScriptableObject
 {
 	[Header("Information")]
-#if UNITY_EDITOR
-	public SceneAsset sceneAsset;
+#if UNITY_EDITOR // See GameSceneSOEditor.cs
+	public UnityEditor.SceneAsset sceneAsset;
 #endif
 	[HideInInspector]
 	public string scenePath;
@@ -22,34 +16,4 @@ public abstract class GameSceneSO : ScriptableObject
 
 	[Header("Sounds")]
 	public AudioClip music;
-
-#if UNITY_EDITOR
-	private SceneAsset prevSceneAsset;
-
-	void ISerializationCallbackReceiver.OnBeforeSerialize()
-	{
-		PopulateScenePath();
-	}
-
-	void ISerializationCallbackReceiver.OnAfterDeserialize()
-	{ }
-
-	private void PopulateScenePath()
-	{
-		if (sceneAsset != null)
-		{
-			// To prevent constant invocation of AssetDatabase API
-			// when this SO is opened in the Inspector.
-			if (prevSceneAsset != sceneAsset)
-			{
-				prevSceneAsset = sceneAsset;
-				scenePath = AssetDatabase.GetAssetPath(sceneAsset);
-			}
-		}
-		else
-		{
-			scenePath = string.Empty;
-		}
-	}
-#endif
 }
